rename to `GenExtension`
